Updated URDF so it plays nice with new controllers
In hydro, simulation functionality is now provided by the
ros_control package and is separate from PR2. With this change comes
new XML.

This commit updates the Husky model to support the new XML. The
existing Husky simulator continue to work, but this prevents problems
when attaching models (e.g. manipualtors) that use the new
ros_control package.